bug fixing

+++ src/main/java/com/toyhe/app/Flotte/Dtos/BoatClasses/BoatClassRegisterRequest.java
@@ -2,7 +2,7 @@ package com.toyhe.app.Flotte.Dtos.BoatClasses;
 
 public record BoatClassRegisterRequest(
         String name,
-        double placesNumber ,
+        int placesNumber ,
         long boatID  ,
         long priceListID
