+++ src/main/java/com/toyhe/app/Flotte/Dtos/BoatClasses/BoatClassResponse.java
@@ -8,7 +8,7 @@ public record BoatClassResponse(
         long boatID ,
         String boatName  ,
         double priceListID,
-        double placeAvailable
+        int placeAvailable
 
 ) {
     public static BoatClassResponse fromBoatClassToDTO(BoatClass boatClass) {
